pyproject.toml
[project]
name = "learn-fast-core"
version = "0.1.0"
description = "LearnFast Core Engine - A hybrid Graph-RAG learning platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "neo4j>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "markitdown[all]>=0.0.1a2",
    "ollama>=0.1.0",
    "pgvector>=0.2.0",
    "python-dotenv>=1.0.0",
    "hypothesis>=6.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "youtube-transcript-api==0.6.3",
    "yt-dlp>=2025.12.8",
    "sqlalchemy>=2.0.46",
    "alembic>=1.18.1",
]
